// src/components/pages/Report/LogAuditTrailUserAccess/DetailPage/components/Detail/Detail.types.ts
export type DetailProps = {
  data?: {
    username?: string;
    access?: string;
    userId?: string;
    status?: string;
    position?: string;
    periode?: string;
    originDivision?: string;
    requesterName?: string;
    destination?: string;
    changeDate?: string;
  };
  isLoading: boolean;
};